# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис сбора и хранения метрик.",
        "title": "Metrics collection API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/": {
            "get": {
                "description": "Генерирует html страницу со списком всех метрик переданных на сервер.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Отдает html со всеми метриками",
                "operationId": "rootHandler",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Позволяет проверить соединение с базой данных.",
                "tags": [
                    "Ping"
                ],
                "summary": "Проверяет соединение с БД",
                "operationId": "connectionPing",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update/": {
            "post": {
                "description": "Сохраняет или обновляет одну метрику из json объекта.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Сохраняет метрику из json",
                "operationId": "storeFromJSON",
                "parameters": [
                    {
                        "description": "Metric data",
                        "name": "metric_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/metric.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update/{type}/{name}/{value}": {
            "post": {
                "description": "Сохраняет или обновляет одну метрику через url запрос.",
                "tags": [
                    "Store"
                ],
                "summary": "Сохраняет метрику из запроса",
                "operationId": "storeFromURL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric value",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "501": {
                        "description": "Unknown metric type",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/updates/": {
            "post": {
                "description": "Сохраняет или обновляет метрики из массива json объектов.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Сохраняет метрики из json",
                "operationId": "storeAllFromJSON",
                "parameters": [
                    {
                        "description": "Metric data",
                        "name": "metric_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/metric.Metrics"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/value/": {
            "post": {
                "description": "Отдает одну метрику согласно имени и типа метрики из json запроса.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Load"
                ],
                "summary": "Отдает метрику из json",
                "operationId": "loadFromJSON",
                "parameters": [
                    {
                        "description": "Metric data",
                        "name": "metric_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/metric.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/metric.Metric"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Metric not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/value/{type}/{name}": {
            "get": {
                "description": "Отдает одну метрику согласно имени и типа из url запроса.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Load"
                ],
                "summary": "Отдает метрику из запроса",
                "operationId": "LoadFromURL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Metric not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "metric.Metric": {
            "type": "object",
            "properties": {
                "delta": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/metric.Type"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "metric.Metrics": {
            "type": "object",
            "properties": {
                "metrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/metric.Metric"
                    }
                }
            }
        },
        "metric.Type": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "Gauge",
                "Counter"
            ]
        }
    }
}